viewModel for DaySummaryView is introduced, to separate the business logic
