+++ Views/MainViews/MealViews/ViewModelDaySummary.swift
@@ -1,0 +1,8 @@
+//
+//  ViewModelDaySummary.swift
+//  BioAge
+//
+//  Created by Gabons on 08/09/25.
+//
+
+import Foundation
